+++ internal/alert/alert.go
@@ -49,11 +49,22 @@ type MWMBAlertGroup struct {
 	TicketSlow  MWMBAlert
 }
 
-type generator bool
+// WindowsRepo knows how to retrieve windows based on the period of time.
+type WindowsRepo interface {
+	GetWindows(ctx context.Context, period time.Duration) (*Windows, error)
+}
 
-// AlertGenerator knows how to generate all the required alerts based on an SLO.
+// Generator knows how to generate all the required alerts based on an SLO.
 // The generated alerts are generic and don't depend on any specific SLO implementation.
-const AlertGenerator = generator(false)
+type Generator struct {
+	windowsRepo WindowsRepo
+}
+
+func NewGenerator(windowsRepo WindowsRepo) Generator {
+	return Generator{
+		windowsRepo: windowsRepo,
+	}
+}
 
 type SLO struct {
 	ID         string
@@ -61,15 +72,9 @@ type SLO struct {
 	Objective  float64
 }
 
-// windowMetadataCatalog are the current supported (and known that work) time windows for alerting.
-var windowMetadataCatalog = map[time.Duration]WindowMetadata{
-	28 * 24 * time.Hour: newMonthWindowMetadata(28 * 24 * time.Hour),
-	30 * 24 * time.Hour: newMonthWindowMetadata(30 * 24 * time.Hour),
-}
-
-func (g generator) GenerateMWMBAlerts(ctx context.Context, slo SLO) (*MWMBAlertGroup, error) {
-	windowMeta, ok := windowMetadataCatalog[slo.TimeWindow]
-	if !ok {
+func (g Generator) GenerateMWMBAlerts(ctx context.Context, slo SLO) (*MWMBAlertGroup, error) {
+	windows, err := g.windowsRepo.GetWindows(ctx, slo.TimeWindow)
+	if err != nil {
 		return nil, fmt.Errorf("the %s SLO period time window is not supported", slo.TimeWindow)
 	}
 
@@ -78,119 +83,37 @@ func (g generator) GenerateMWMBAlerts(ctx context.Context, slo SLO) (*MWMBAlertG
 	group := MWMBAlertGroup{
 		PageQuick: MWMBAlert{
 			ID:             fmt.Sprintf("%s-page-quick", slo.ID),
-			ShortWindow:    windowMeta.WindowPageQuickShort,
-			LongWindow:     windowMeta.WindowPageQuickLong,
-			BurnRateFactor: windowMeta.GetSpeedPageQuick(),
+			ShortWindow:    windows.PageQuick.ShortWindow,
+			LongWindow:     windows.PageQuick.LongWindow,
+			BurnRateFactor: windows.GetSpeedPageQuick(),
 			ErrorBudget:    errorBudget,
 			Severity:       PageAlertSeverity,
 		},
 		PageSlow: MWMBAlert{
 			ID:             fmt.Sprintf("%s-page-slow", slo.ID),
-			ShortWindow:    windowMeta.WindowPageSlowShort,
-			LongWindow:     windowMeta.WindowPageSlowLong,
-			BurnRateFactor: windowMeta.GetSpeedPageSlow(),
+			ShortWindow:    windows.PageSlow.ShortWindow,
+			LongWindow:     windows.PageSlow.LongWindow,
+			BurnRateFactor: windows.GetSpeedPageSlow(),
 			ErrorBudget:    errorBudget,
 			Severity:       PageAlertSeverity,
 		},
 		TicketQuick: MWMBAlert{
 			ID:             fmt.Sprintf("%s-ticket-quick", slo.ID),
-			ShortWindow:    windowMeta.WindowTicketQuickShort,
-			LongWindow:     windowMeta.WindowTicketQuickLong,
-			BurnRateFactor: windowMeta.GetSpeedTicketQuick(),
+			ShortWindow:    windows.TicketQuick.ShortWindow,
+			LongWindow:     windows.TicketQuick.LongWindow,
+			BurnRateFactor: windows.GetSpeedTicketQuick(),
 			ErrorBudget:    errorBudget,
 			Severity:       TicketAlertSeverity,
 		},
 		TicketSlow: MWMBAlert{
 			ID:             fmt.Sprintf("%s-ticket-slow", slo.ID),
-			ShortWindow:    windowMeta.WindowTicketSlowShort,
-			LongWindow:     windowMeta.WindowTicketSlowLong,
-			BurnRateFactor: windowMeta.GetSpeedTicketSlow(),
+			ShortWindow:    windows.TicketSlow.ShortWindow,
+			LongWindow:     windows.TicketSlow.LongWindow,
+			BurnRateFactor: windows.GetSpeedTicketSlow(),
 			ErrorBudget:    errorBudget,
 			Severity:       TicketAlertSeverity,
 		},
 	}
 
 	return &group, nil
-}
-
-// WindowMetadata has the information required to calculate SLOs.
-type WindowMetadata struct {
-	WindowPeriod time.Duration
-
-	// Alerting required windows.
-	// Its a matrix of values with:
-	// - Alert severity: ["page", "ticket"].
-	// - Measure period: ["long", "short"].
-	WindowPageQuickShort   time.Duration
-	WindowPageQuickLong    time.Duration
-	WindowPageSlowShort    time.Duration
-	WindowPageSlowLong     time.Duration
-	WindowTicketQuickShort time.Duration
-	WindowTicketQuickLong  time.Duration
-	WindowTicketSlowShort  time.Duration
-	WindowTicketSlowLong   time.Duration
-
-	// Error budget percent consumed for a full time window.
-	// Google gives us some defaults in its SRE workbook that work correctly most of the times:
-	// - Page quick:   2%
-	// - Page slow:    5%
-	// - Ticket quick: 10%
-	// - Ticket slow:  10%
-	ErrorBudgetPercPageQuick   float64
-	ErrorBudgetPercPageSlow    float64
-	ErrorBudgetPercTicketQuick float64
-	ErrorBudgetPercTicketSlow  float64
-}
-
-// Error budget speeds based on a full time window, however once we have the factor (speed)
-// the value can be used with any time window.
-func (w WindowMetadata) GetSpeedPageQuick() float64 {
-	return w.getBurnRateFactor(w.WindowPeriod, w.ErrorBudgetPercPageQuick, w.WindowPageQuickLong)
-}
-func (w WindowMetadata) GetSpeedPageSlow() float64 {
-	return w.getBurnRateFactor(w.WindowPeriod, w.ErrorBudgetPercPageSlow, w.WindowPageSlowLong)
-}
-func (w WindowMetadata) GetSpeedTicketQuick() float64 {
-	return w.getBurnRateFactor(w.WindowPeriod, w.ErrorBudgetPercTicketQuick, w.WindowTicketQuickLong)
-}
-func (w WindowMetadata) GetSpeedTicketSlow() float64 {
-	return w.getBurnRateFactor(w.WindowPeriod, w.ErrorBudgetPercTicketSlow, w.WindowTicketSlowLong)
-}
-
-// getBurnRateFactor calculates the burnRateFactor (speed) needed to consume all the error budget available percent
-// in a specific time window taking into account the total time window.
-func (w WindowMetadata) getBurnRateFactor(totalWindow time.Duration, errorBudgetPercent float64, consumptionWindow time.Duration) float64 {
-	// First get the total hours required to consume the % of the error budget in the total window.
-	hoursRequiredConsumption := errorBudgetPercent * totalWindow.Hours() / 100
-
-	// Now calculate how much is the factor required for the hours consumption, in case we would need to use
-	// a different time window (e.g: hours required: 36h, if we want to do it in 6h: would be `x6`).
-	speed := hoursRequiredConsumption / consumptionWindow.Hours()
-
-	return speed
-}
-
-// newMonthWindowMetadata returns a common and safe approximate month window metadata. Normally this works well
-// with 4-5 weeks time windows like 28 day and 30 day.
-// Is the most common kind of SLO based window metadata.
-//
-// Numbers obtained from https://sre.google/workbook/alerting-on-slos/#recommended_parameters_for_an_slo_based_a.
-func newMonthWindowMetadata(windowPeriod time.Duration) WindowMetadata {
-	return WindowMetadata{
-		WindowPeriod: windowPeriod,
-
-		WindowPageQuickShort:   5 * time.Minute,
-		WindowPageQuickLong:    1 * time.Hour,
-		WindowPageSlowShort:    30 * time.Minute,
-		WindowPageSlowLong:     6 * time.Hour,
-		WindowTicketQuickShort: 2 * time.Hour,
-		WindowTicketQuickLong:  1 * 24 * time.Hour,
-		WindowTicketSlowShort:  6 * time.Hour,
-		WindowTicketSlowLong:   3 * 24 * time.Hour,
-
-		ErrorBudgetPercPageQuick:   2,
-		ErrorBudgetPercPageSlow:    5,
-		ErrorBudgetPercTicketQuick: 10,
-		ErrorBudgetPercTicketSlow:  10,
-	}
 }