+++ test/integration/k8scontroller/k8scontroller_test.go
@@ -3,6 +3,7 @@ package k8scontroller_test
 import (
 	"context"
 	"fmt"
+	"strings"
 	"testing"
 	"time"
 
@@ -47,11 +48,12 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 
 	// Tests.
 	tests := map[string]struct {
-		windowDays int
-		exec       func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients)
+		customSLOPeriods bool
+		sloPeriod        string
+		exec             func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients)
 	}{
 		"Having SLOs as a CRD should generate Prometheus operator CRD.": {
-			windowDays: 30,
+			sloPeriod: "30d",
 			exec: func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients) {
 				// Prepare our SLO on Kubernetes.
 				SLOs := getBasePrometheusServiceLevel()
@@ -59,7 +61,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 				require.NoError(t, err)
 
 				// Wait to be sure the controller had time for handling.
-				time.Sleep(250 * time.Millisecond)
+				time.Sleep(500 * time.Millisecond)
 
 				// Check.
 				expRule := getBasePromOpPrometheusRule(version)
@@ -74,7 +76,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 		},
 
 		"Having SLOs with custom time windows (28 day) should generate Prometheus operator CRD.": {
-			windowDays: 28,
+			sloPeriod: "28d",
 			exec: func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients) {
 				// Prepare our SLO on Kubernetes.
 				SLOs := getBasePrometheusServiceLevel()
@@ -82,7 +84,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 				require.NoError(t, err)
 
 				// Wait to be sure the controller had time for handling.
-				time.Sleep(250 * time.Millisecond)
+				time.Sleep(500 * time.Millisecond)
 
 				// Check.
 				expRule := getBase28DayPromOpPrometheusRule(version)
@@ -97,7 +99,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 		},
 
 		"Having SLOs with plugins as a CRD should generate Prometheus operator CRD.": {
-			windowDays: 30,
+			sloPeriod: "30d",
 			exec: func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients) {
 				// Prepare our SLO on Kubernetes with plugin based SLO.
 				SLOs := getPluginPrometheusServiceLevel()
@@ -105,7 +107,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 				require.NoError(t, err)
 
 				// Wait to be sure the controller had time for handling.
-				time.Sleep(250 * time.Millisecond)
+				time.Sleep(500 * time.Millisecond)
 
 				// Check.
 				expRule := getPluginPromOpPrometheusRule(version)
@@ -120,7 +122,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 		},
 
 		"Having SLOs as a CRD should set the status as correct on the CRD.": {
-			windowDays: 30,
+			sloPeriod: "30d",
 			exec: func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients) {
 				// Prepare our SLO on Kubernetes.
 				SLOs := getBasePrometheusServiceLevel()
@@ -128,7 +130,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 				require.NoError(t, err)
 
 				// Wait to be sure the controller had time for handling.
-				time.Sleep(250 * time.Millisecond)
+				time.Sleep(500 * time.Millisecond)
 
 				// Check.
 				gotSLOs, err := kubeClis.Sloth.SlothV1().PrometheusServiceLevels(ns).Get(ctx, SLOs.Name, metav1.GetOptions{})
@@ -147,7 +149,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 		},
 
 		"Having wrong SLOs as a CRD should set the status failed on the CRD.": {
-			windowDays: 30,
+			sloPeriod: "30d",
 			exec: func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients) {
 				// Prepare our wrong SLO on Kubernetes.
 				SLOs := getBasePrometheusServiceLevel()
@@ -156,7 +158,7 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 				require.NoError(t, err)
 
 				// Wait to be sure the controller had time for handling.
-				time.Sleep(250 * time.Millisecond)
+				time.Sleep(500 * time.Millisecond)
 
 				// Check.
 				gotSLOs, err := kubeClis.Sloth.SlothV1().PrometheusServiceLevels(ns).Get(ctx, SLOs.Name, metav1.GetOptions{})
@@ -171,6 +173,30 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 				gotSLOs.Status.LastPromOpRulesSuccessfulGenerated = nil // Remove variations.
 
 				assert.Equal(t, expStatus, gotSLOs.Status)
+			},
+		},
+
+		"Having SLOs with custom external time windows (7 day) should generate Prometheus operator CRD.": {
+			customSLOPeriods: true,
+			sloPeriod:        "7d",
+			exec: func(ctx context.Context, t *testing.T, ns string, kubeClis *k8scontroller.KubeClients) {
+				// Prepare our SLO on Kubernetes.
+				SLOs := getBasePrometheusServiceLevel()
+				_, err := kubeClis.Sloth.SlothV1().PrometheusServiceLevels(ns).Create(ctx, SLOs, metav1.CreateOptions{})
+				require.NoError(t, err)
+
+				// Wait to be sure the controller had time for handling.
+				time.Sleep(500 * time.Millisecond)
+
+				// Check.
+				expRule := getBase7DayPromOpPrometheusRule(version)
+				expRule.Namespace = ns
+
+				gotRule, err := kubeClis.Monitoring.MonitoringV1().PrometheusRules(ns).Get(ctx, expRule.Name, metav1.GetOptions{})
+				gotRule = sanitizePrometheusRule(gotRule) // Remove variations.
+				require.NoError(t, err)
+
+				assert.Equal(t, expRule, gotRule)
 			},
 		},
 	}
@@ -195,8 +221,20 @@ func TestKubernetesControllerPromOperatorGenerate(t *testing.T) {
 
 			// Run controller in background.
 			go func() {
+				// Prepare args.
+				args := []string{
+					"--metrics-listen-addr=:0",
+					"--hot-reload-addr=:0",
+					"--sli-plugins-path=./",
+					fmt.Sprintf("--namespace=%s", ns),
+					fmt.Sprintf("--default-slo-period=%s", test.sloPeriod),
+				}
+				if test.customSLOPeriods {
+					args = append(args, "--slo-period-windows-path=./windows")
+				}
+
 				// Listen on `:0` and isolate per namespace so we can run in tests parallel safely.
-				_, _, _ = k8scontroller.RunSlothController(ctx, config, ns, fmt.Sprintf("--metrics-listen-addr=:0 --hot-reload-addr=:0 --namespace=%s --window-days=%d", ns, test.windowDays))
+				_, _, _ = k8scontroller.RunSlothController(ctx, config, ns, strings.Join(args, " "))
 			}()
 
 			// Execute test.